+++ src/analyze.test.ts
@@ -13,7 +13,7 @@ import {
 } from "./analyze";
 import { setCodeQL } from "./codeql";
 import { Feature } from "./feature-flags";
-import { Language } from "./languages";
+import { KnownLanguage } from "./languages";
 import { getRunnerLogger } from "./logging";
 import {
   setupTests,
@@ -41,7 +41,7 @@ test("status report fields", async (t) => {
     const threadsFlag = "";
     sinon.stub(uploadLib, "validateSarifFileSchema");
 
-    for (const language of Object.values(Language)) {
+    for (const language of Object.values(KnownLanguage)) {
       setCodeQL({
         databaseRunQueries: async () => {},
         packDownload: async () => ({ packs: [] }),
@@ -106,7 +106,6 @@ test("status report fields", async (t) => {
         memoryFlag,
         addSnippetsFlag,
         threadsFlag,
-        "brutal",
         undefined,
         undefined,
         config,
@@ -331,13 +330,13 @@ test("getDiffRanges: malformed thunk header", async (t) => {
 test("resolveQuerySuiteAlias", (t) => {
   // default query suite names should resolve to something language-specific ending in `.qls`.
   for (const suite of defaultSuites) {
-    const resolved = resolveQuerySuiteAlias(Language.go, suite);
+    const resolved = resolveQuerySuiteAlias(KnownLanguage.go, suite);
     t.assert(
       resolved.endsWith(".qls"),
       "Resolved default suite doesn't end in .qls",
     );
     t.assert(
-      resolved.indexOf(Language.go) >= 0,
+      resolved.indexOf(KnownLanguage.go) >= 0,
       "Resolved default suite doesn't contain language name",
     );
   }
@@ -346,6 +345,6 @@ test("resolveQuerySuiteAlias", (t) => {
   const names = ["foo", "bar", "codeql/go-queries@1.0"];
 
   for (const name of names) {
-    t.deepEqual(resolveQuerySuiteAlias(Language.go, name), name);
+    t.deepEqual(resolveQuerySuiteAlias(KnownLanguage.go, name), name);
   }
 });