+++ src/upload-lib.test.ts
@@ -3,9 +3,8 @@ import * as path from "path";
 
 import test from "ava";
 
-import { Feature } from "./feature-flags";
 import { getRunnerLogger, Logger } from "./logging";
-import { createFeatures, setupTests } from "./testing-utils";
+import { setupTests } from "./testing-utils";
 import * as uploadLib from "./upload-lib";
 import { GitHubVariant, initializeEnvironment, withTmpDir } from "./util";
 
@@ -457,11 +456,10 @@ test("shouldShowCombineSarifFilesDeprecationWarning when environment variable is
   );
 });
 
-test("throwIfCombineSarifFilesDisabled when on dotcom with feature flag", async (t) => {
+test("throwIfCombineSarifFilesDisabled when on dotcom", async (t) => {
   await t.throwsAsync(
     uploadLib.throwIfCombineSarifFilesDisabled(
       [createMockSarif("abc", "def"), createMockSarif("abc", "def")],
-      createFeatures([Feature.DisableCombineSarifFiles]),
       {
         type: GitHubVariant.DOTCOM,
       },
@@ -473,23 +471,10 @@ test("throwIfCombineSarifFilesDisabled when on dotcom with feature flag", async 
   );
 });
 
-test("throwIfCombineSarifFilesDisabled when on dotcom without feature flag", async (t) => {
-  await t.notThrowsAsync(
-    uploadLib.throwIfCombineSarifFilesDisabled(
-      [createMockSarif("abc", "def"), createMockSarif("abc", "def")],
-      createFeatures([]),
-      {
-        type: GitHubVariant.DOTCOM,
-      },
-    ),
-  );
-});
-
 test("throwIfCombineSarifFilesDisabled when on GHES 3.13", async (t) => {
   await t.notThrowsAsync(
     uploadLib.throwIfCombineSarifFilesDisabled(
       [createMockSarif("abc", "def"), createMockSarif("abc", "def")],
-      createFeatures([]),
       {
         type: GitHubVariant.GHES,
         version: "3.13.2",
@@ -502,7 +487,6 @@ test("throwIfCombineSarifFilesDisabled when on GHES 3.14", async (t) => {
   await t.notThrowsAsync(
     uploadLib.throwIfCombineSarifFilesDisabled(
       [createMockSarif("abc", "def"), createMockSarif("abc", "def")],
-      createFeatures([]),
       {
         type: GitHubVariant.GHES,
         version: "3.14.0",
@@ -515,7 +499,6 @@ test("throwIfCombineSarifFilesDisabled when on GHES 3.17", async (t) => {
   await t.notThrowsAsync(
     uploadLib.throwIfCombineSarifFilesDisabled(
       [createMockSarif("abc", "def"), createMockSarif("abc", "def")],
-      createFeatures([]),
       {
         type: GitHubVariant.GHES,
         version: "3.17.0",
@@ -528,7 +511,6 @@ test("throwIfCombineSarifFilesDisabled when on GHES 3.18 pre", async (t) => {
   await t.throwsAsync(
     uploadLib.throwIfCombineSarifFilesDisabled(
       [createMockSarif("abc", "def"), createMockSarif("abc", "def")],
-      createFeatures([]),
       {
         type: GitHubVariant.GHES,
         version: "3.18.0.pre1",
@@ -545,7 +527,6 @@ test("throwIfCombineSarifFilesDisabled when on GHES 3.18 alpha", async (t) => {
   await t.throwsAsync(
     uploadLib.throwIfCombineSarifFilesDisabled(
       [createMockSarif("abc", "def"), createMockSarif("abc", "def")],
-      createFeatures([]),
       {
         type: GitHubVariant.GHES,
         version: "3.18.0-alpha.1",
@@ -562,7 +543,6 @@ test("throwIfCombineSarifFilesDisabled when on GHES 3.18", async (t) => {
   await t.throwsAsync(
     uploadLib.throwIfCombineSarifFilesDisabled(
       [createMockSarif("abc", "def"), createMockSarif("abc", "def")],
-      createFeatures([]),
       {
         type: GitHubVariant.GHES,
         version: "3.18.0",
@@ -579,7 +559,6 @@ test("throwIfCombineSarifFilesDisabled with an invalid GHES version", async (t) 
   await t.notThrowsAsync(
     uploadLib.throwIfCombineSarifFilesDisabled(
       [createMockSarif("abc", "def"), createMockSarif("abc", "def")],
-      createFeatures([]),
       {
         type: GitHubVariant.GHES,
         version: "foobar",
@@ -592,7 +571,6 @@ test("throwIfCombineSarifFilesDisabled with only 1 run", async (t) => {
   await t.notThrowsAsync(
     uploadLib.throwIfCombineSarifFilesDisabled(
       [createMockSarif("abc", "def")],
-      createFeatures([Feature.DisableCombineSarifFiles]),
       {
         type: GitHubVariant.DOTCOM,
       },
@@ -604,7 +582,6 @@ test("throwIfCombineSarifFilesDisabled with distinct categories", async (t) => {
   await t.notThrowsAsync(
     uploadLib.throwIfCombineSarifFilesDisabled(
       [createMockSarif("abc", "def"), createMockSarif("def", "def")],
-      createFeatures([Feature.DisableCombineSarifFiles]),
       {
         type: GitHubVariant.DOTCOM,
       },
@@ -616,7 +593,6 @@ test("throwIfCombineSarifFilesDisabled with distinct tools", async (t) => {
   await t.notThrowsAsync(
     uploadLib.throwIfCombineSarifFilesDisabled(
       [createMockSarif("abc", "abc"), createMockSarif("abc", "def")],
-      createFeatures([Feature.DisableCombineSarifFiles]),
       {
         type: GitHubVariant.DOTCOM,
       },